% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{compute_anomaly}
\alias{compute_anomaly}
\title{compute_anomaly}
\usage{
compute_anomaly(pathLength, dataSize)
}
\arguments{
\item{pathLength}{pathLength}

\item{dataSize}{dataSize}
}
\description{
compute_anomaly
}
