% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{depth_corrected}
\alias{depth_corrected}
\title{depth_corrected}
\usage{
depth_corrected(object, newdata, ...)
}
\arguments{
\item{object}{isolation forest model}

\item{newdata}{dataframe to predict}

\item{...}{Arguments to be passed to future.apply::future_lapply}
}
\description{
depth_corrected
}
