% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{depth_terminalNodes}
\alias{depth_terminalNodes}
\title{depth_terminalNodes}
\usage{
depth_terminalNodes(treelike)
}
\arguments{
\item{treelike}{A single ranger tree extracted from 'ranger::treeInfo'}
}
\description{
depth_terminalNodes
}
