% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_integerish}
\alias{is_integerish}
\title{Check for a single integer}
\usage{
is_integerish(x)
}
\arguments{
\item{x}{input}
}
\value{
TRUE or FALSE
}
\description{
for a single integer
}
\examples{
\dontrun{is_integerish(1)}
}
