% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_solitude.R
\docType{package}
\name{solitude}
\alias{solitude}
\alias{solitude-package}
\title{An Implementation of Isolation Forest}
\description{
Isolation forest is anomaly detection method introduced by the paper Isolation based Anomaly Detection (Liu, Ting and Zhou <doi:10.1145/2133360.2133363>)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talegari/solitude}
  \item Report bugs at \url{https://github.com/talegari/solitude/issues}
}

}
\author{
Srikanth Komala Sheshachala
}
