% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_nodes.R
\name{terminalNodesDepth}
\alias{terminalNodesDepth}
\title{Depth of each terminal node of all trees in a ranger model}
\usage{
terminalNodesDepth(model)
}
\arguments{
\item{model}{A ranger model}
}
\value{
A tibble with three columns: 'id_tree', 'id_node', 'depth'.
}
\description{
Depth of each terminal node of all trees in a ranger model is
  returned as a three column tibble with column names: 'id_tree',
  'id_node', 'depth'. Note that root node has the node_id = 0.
}
\details{
This function may be parallelized using a future backend.
}
\examples{
rf = ranger::ranger(Species ~ ., data = iris, num.trees = 100)
terminalNodesDepth(rf)
}
