% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{anomaly_score}
\alias{anomaly_score}
\title{anomaly_score}
\usage{
anomaly_score(object, newdata, aggregator = "median")
}
\arguments{
\item{object}{isolation forest model}

\item{newdata}{dataframe to predict}

\item{aggregator}{aggregator}
}
\description{
anomaly_score
}
