% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolation_forest.R
\name{isolationForest}
\alias{isolationForest}
\alias{isolation_forest}
\title{Grow an isolation forest}
\usage{
isolationForest(dataset, ...)
}
\arguments{
\item{dataset}{Dataframe with column names}

\item{...}{Optional arguments passed to \code{\link[ranger]{ranger}}}
}
\value{
An object(list) of class 'solitude'. The isolationforest is in the
  'forest' slot
}
\description{
Implements isolation forest  from Isolation-Based Anomaly
  Detection (<doi:10.1145/2133360.2133363>)
}
\details{
This function internally uses \code{\link[ranger]{ranger}}.
  Arguments may be passed to it except these: dependent.variable.name, data,
  mtry, min.node.size, splitrule, num.random.splits,
  respect.unordered.factors. Pass 'seed' argument for reproducibility
}
\examples{
mo <- isolationForest(iris, seed = 1)
}
